Update thread.c
This program creates NUM_THREADS threads and prints their respective user-assigned thread id. The first thing to notice is the call to pthread_create() in the main function. The syntax of the third and fourth argument are particularly important. Notice that the thr_func is the name of the thread function, while the fourth argument is the argument passed to said function. Here we are passing a thread function argument that we created as a thread_data_t struct. Of course, you can pass simple data types as pointers if that is all that is needed, or NULL if no arguments are required. However, it is good practice to be able to pass arguments of arbitrary type and size, and is thus illustrated for this purpose.

A few things to mention:

Make sure you check the return values for all important functions.
The second argument to pthread_create() is NULL indicating to create a thread with default attributes. The defaults vary depend upon the system and pthread implementation.
Notice that we have broken apart the pthread_join() from the pthread_create(). Why is it that you should not integrate the pthread_join() in to the thread creation loop?
Although not explicitly required to call pthread_exit() at the end of the thread function, it is good practice to do so, as you may have the need to return some arbitrary data back to the caller via pthread_join().